+++ src/frontend/mod.rs
@@ -1,4 +1,6 @@
 //! CorTeX Frontend capabilities - reports, cache, captchas and more
 pub mod cached;
 pub mod captcha;
+pub mod concerns;
+pub mod helpers;
 pub mod params;
